Move ErrorWidget logic to Errors panel.
